Disable plugin if wrench item can't be properly set
I really like initializer blocks because I like being able to see what I need to know about a variable in the same place it was declared. However, moving the item meta changes to the onEnable method allows me to explicitly control how to handle failure. This is only relevant if the api or implementation changes because right now carrot on a stick ItemMeta is Damageable.